requirements.txt
fastapi
uvicorn[standard]
playwright
openai
pydantic
gunicorn
python-dotenv